Adding javascript confirm before deletion

+++ admin/functions.php
@@ -129,7 +129,7 @@
                     <td>{$post_status}</td>
                     <td>{$post_comment_count}</td>
                     <td><a href='posts.php?source=edit_post&post_id={$post_id}'>Edit</a></td>
-                    <td><a href='posts.php?delete={$post_id}'>Delete</a></td>
+                    <td><a onclick=\"javascript: return confirm('Are you sure you want to delete?')\" href='posts.php?delete={$post_id}'>Delete</a></td>
                 </tr>";
         }
     }
